pyproject.toml
[project]
name = "panodac"
version = "0.1.0"
description = "Plug-and-play metric depth estimation for any camera (perspective, fisheye, 360°). Based on Depth Any Camera (CVPR 2025)."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
    { name = "Yuliang Guo" },
    { name = "Sparsh Garg" },
]
keywords = ["depth-estimation", "monocular-depth", "panorama", "fisheye", "360", "computer-vision", "pytorch"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy<2.0.0",
    "opencv-python>=4.5.0",
    "pillow>=9.0.0",
    "einops>=0.6.0",
    "timm>=0.9.0",
    "huggingface-hub>=0.20.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yuliangguo/depth_any_camera"
Repository = "https://github.com/YOUR_USERNAME/panodac"
Documentation = "https://github.com/YOUR_USERNAME/panodac#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/panodac"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
